refactor(app/controllers):  add modules on controllers
